+++ lib/rubocop/cop/style/one_line_conditional.rb
@@ -20,6 +20,27 @@ module RuboCop
 
           add_offense(node, :expression, format(MSG, condition))
         end
+
+        def autocorrect(node)
+          lambda do |corrector|
+            corrector.replace(node.source_range, replacement(node))
+          end
+        end
+
+        def replacement(node)
+          cond, body, else_clause = *node
+          ternary = "#{cond.source} ? #{body.source} : #{else_clause.source}"
+
+          return ternary unless node.parent
+          return "(#{ternary})" if [:and, :or].include?(node.parent.type)
+
+          if node.parent.send_type?
+            _receiver, method_name, = *node.parent
+            return "(#{ternary})" if operator?(method_name)
+          end
+
+          ternary
+        end
       end
     end
   end